Create detect_mobile_device.php
This function checks whether the user is accessing the website from a mobile device or not based on the user agent string. It uses a series of regular expressions to match known mobile device signatures (including Android, iPhone, Windows Phone, Blackberry, and others). If the device is identified as a mobile device, it returns true, otherwise, it returns false.